refactor: moving weaver logging to extensions
