+++ Assets/Mirage/Weaver/Mirage.CecilExtensions/Logging.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6f45e6a35a4f495449af830d83af5573
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
