+++ Assets/Mirage/Weaver/Mirage.CecilExtensions/Logging/IWeaverLogger.cs
@@ -1,7 +1,9 @@
+using System.Collections.Generic;
 using Mono.Cecil;
 using Mono.Cecil.Cil;
+using Unity.CompilationPipeline.Common.Diagnostics;
 
-namespace Mirage.Weaver
+namespace Mirage.CodeGen
 {
     // todo rename this to IWeaverErrors because it isn't really a logger
     public interface IWeaverLogger
@@ -20,26 +22,7 @@ namespace Mirage.Weaver
         void Warning(string message, MemberReference mr);
         void Warning(string message, MemberReference mr, SequencePoint sequencePoint);
         void Warning(string message, MethodDefinition md);
-    }
 
-    internal static class WeaverLoggerExtensions
-    {
-        public static void Error(this IWeaverLogger logger, WeaverException exception)
-        {
-            var message = exception.Message;
-            if (logger.EnableTrace)
-                message += "\n" + exception.ToString();
-
-            logger.Error(message, exception.MemberReference, exception.SequencePoint);
-        }
-
-        public static void Error(this IWeaverLogger logger, WeaverException exception, SequencePoint sequencePoint)
-        {
-            var message = exception.Message;
-            if (logger.EnableTrace)
-                message += "\n" + exception.ToString();
-
-            logger.Error(message, exception.MemberReference, sequencePoint);
-        }
+        List<DiagnosticMessage> GetDiagnostics();
     }
 }
