+++ Assets/Mirage/Weaver/Mirage.CecilExtensions/Logging/LoggerExtensions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d2c08fb94322cbb489fc2dfd57e2589c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
