+++ Assets/Mirage/Weaver/Mirage.CecilExtensions/Logging/WeaverException.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2c08ab263a8f0dd4995771d6a180f6de
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
