+++ Assets/Mirage/Weaver/Mirage.CecilExtensions/Logging/WeaverLogger.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ec36b2acb3f2954488ef1e7d0c07a864
+guid: 50969081e34a483499cfdaa9a13e602c
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
