+++ Assets/Mirage/Weaver/WeaverExceptions.cs
@@ -1,35 +1,10 @@
-using System;
 using System.Linq;
+using Mirage.CodeGen;
 using Mono.Cecil;
 using Mono.Cecil.Cil;
 
 namespace Mirage.Weaver
 {
-    /// <summary>
-    /// Weaver Exception should be thrown when there is a problem with the users code that they should fix
-    /// <para>
-    ///     For example, if the user uses an unsupported type a WeaverException should be thrown with a
-    ///     message explaining the problem, and the MemberReference to help the user find the issue
-    /// </para>
-    /// <para>
-    ///     For Exception that are internally to weaver (eg weaver didn't work right) and normal Exception should be thrown
-    /// </para>
-    /// </summary>
-    // should be caught within weaver and returned to user using DiagnosticMessage
-#pragma warning disable S3871 // Exception types should be "public"
-    internal abstract class WeaverException : Exception
-#pragma warning restore S3871 // Exception types should be "public"
-    {
-        public readonly SequencePoint SequencePoint;
-        public readonly MemberReference MemberReference;
-
-        protected WeaverException(string message, MemberReference memberReference, SequencePoint sequencePoint) : base(message)
-        {
-            SequencePoint = sequencePoint;
-            MemberReference = memberReference;
-        }
-    }
-
     /// <summary>
     /// Thrown when can't generate read or write for a type
     /// </summary>
